+++ tests/testthat/test-conditions.R
@@ -168,3 +168,14 @@ test_that("restarting() handlers pass along all requested arguments", {
   }
   with_restarts(fn(), rst_foo = rst_foo)
 })
+
+test_that("cnd_signal() returns NULL invisibly", {
+  expect_identical(withVisible(cnd_signal("foo")), withVisible(invisible(NULL)))
+})
+
+test_that("cnd_signal() accepts character vectors (#195)", {
+  expect <- inplace(function(cnd) {
+    expect_identical(class(cnd), c("mufflable", "foo", "bar", "condition"))
+  })
+  with_handlers(cnd_signal(c("foo", "bar")), foo = expect)
+})