+++ NEWS.md
@@ -11,6 +11,11 @@
   a call depth as `call` arguments. This allows plucking a call from
   further up the call stack (#30).
 
+* `cnd_signal()` now returns invisibly.
+
+* `cnd_signal()` and `cnd_abort()` now accept character vectors to
+  create typed conditions with several S3 subclasses.
+
 * New `env_set()` function to set a value in an environment or a
   scope. If the `create` argument is `FALSE`, it only overwrites
   existing bindings and issues an error otherwise (#162).